fix constraint violation and send conflict status 409 while deleting location

+++ src/main/java/com/resume/web/controller/LocationController.java
@@ -1,21 +1,24 @@
 package com.resume.web.controller;
 
+import com.resume.Services.IExperienceService;
+import com.resume.Services.IFormationService;
+import com.resume.Services.ILocationService;
 import com.resume.converter.ConverterHelper;
 import com.resume.dco.LocationDco;
+import com.resume.dto.ExperienceDto;
+import com.resume.dto.FormationDto;
 import com.resume.dto.LocationDto;
-import com.resume.model.Competence;
-import com.resume.model.Location;
-import com.resume.repository.LocationRepository;
 import com.resume.web.exceptions.NoContentException;
 import com.resume.web.exceptions.UnexpectedCompetenceException;
 import com.resume.web.exceptions.UnexpectedLocationException;
-import com.resume.web.serviceImpl.LocationServiceImpl;
 import io.swagger.annotations.ApiOperation;
 import org.modelmapper.ModelMapper;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import javax.validation.Valid;
 import java.util.List;
@@ -25,15 +28,19 @@ public class LocationController {
 
     public static final Logger logger = LoggerFactory.getLogger(LocationController.class);
 
-    private LocationServiceImpl locationService;
+    private final ILocationService locationService;
 
-    private LocationRepository locationRepository;
+    private final IFormationService formationService;
 
-    private ModelMapper modelMapper;
+    private final IExperienceService experienceService;
 
-    public LocationController(LocationServiceImpl locationService, LocationRepository locationRepository, ModelMapper modelMapper) {
+    private final ModelMapper modelMapper;
+
+    public LocationController(ILocationService locationService, IFormationService formationService,
+                              IExperienceService experienceService, ModelMapper modelMapper) {
         this.locationService = locationService;
-        this.locationRepository = locationRepository;
+        this.formationService = formationService;
+        this.experienceService = experienceService;
         this.modelMapper = modelMapper;
     }
 
@@ -71,30 +78,48 @@ public class LocationController {
     public ResponseEntity<LocationDto> put(@Valid @RequestBody LocationDco locationDco) {
         logger.info("Updating location with id {}", locationDco.getLocationId());
 
-        Location currentLocation = locationRepository.findByLocationId(locationDco.getLocationId());
+        LocationDto currentLocation = locationService.putLocation(locationDco);
 
         if (currentLocation == null) {
             logger.error("Unable to update. Location with id {} not found.", locationDco.getLocationId());
-            throw new UnexpectedLocationException("Unable to upate. Location with id " + locationDco.getLocationId() + " not found.");
+            throw new UnexpectedLocationException("Unable to update. Location with id " + locationDco.getLocationId() + " not found.");
         }
 
-        currentLocation.setCity(locationDco.getCity());
-        currentLocation.setCountry(locationDco.getCountry());
-        currentLocation.setNumber(locationDco.getNumber());
-        currentLocation.setStreet(locationDco.getStreet());
-        currentLocation.setZipCode(locationDco.getZipCode());
-
-        return ResponseEntity.ok().body(locationService.putLocation(currentLocation));
+        return ResponseEntity.ok().body(currentLocation);
     }
 
     @CrossOrigin()
     @ApiOperation(value = "Delete a location")
     @DeleteMapping(value = "/Locations/{locationId}")
-    public ResponseEntity<Location> delete(@PathVariable Integer locationId) {
+    public ResponseEntity<LocationDto> delete(@PathVariable Integer locationId) {
         logger.info("Deleting location with id {}", locationId);
-        Location locationToDelete = locationRepository.findByLocationId(locationId);
-        Integer deletedLocation  = locationService.deleteLocation(locationId);
+        LocationDto locationToDelete = locationService.getLocationbyId(locationId);
 
+        if(locationToDelete == null) {
+            throw new UnexpectedCompetenceException("Unable to delete. Location with id " + locationId + " not found.");
+        }
+
+        List<FormationDto> formationLinkedToLocation = formationService.getFormationLinkedToLocation(locationToDelete.getLocationId());
+        if (formationLinkedToLocation != null && !formationLinkedToLocation.isEmpty())  {
+            String dependantFormations = formationLinkedToLocation.stream()
+                                        .map(FormationDto::toString)
+                                        .reduce("\n", String::concat);
+            throw new ResponseStatusException(HttpStatus.CONFLICT,
+                    "There are some formations still linked to the location you want to delete.\n" +
+                            "Please delete the link to the following elements and try again later:\n" + dependantFormations);
+        }
+
+        List<ExperienceDto> experienceLinkedToLocation = experienceService.getExperienceLinkedToLocation(locationToDelete.getLocationId());
+        if (experienceLinkedToLocation != null && !experienceLinkedToLocation.isEmpty())  {
+            String dependantExperiences = experienceLinkedToLocation.stream()
+                    .map(ExperienceDto::toString)
+                    .reduce("\n", String::concat);
+            throw new ResponseStatusException(HttpStatus.CONFLICT,
+                    "There are some experiences still linked to the location you want to delete.\n" +
+                            "Please delete the link to the following elements and try again later:\n" + dependantExperiences);
+        }
+
+        Integer deletedLocation  = locationService.deleteLocation(locationId);
         if(deletedLocation == 0) {
             throw new UnexpectedCompetenceException("Unable to delete. Location with id " + locationId + " not found.");
         }
